src/terminator/TerminatorServer.java: Allow the server to accept a workingDirectory argument.
bin/terminator: Always pass the workingDirectory to the server.  Initialize it from the current directory (rather than the home directory) if the user doesn't specify it on the command line.  I've done it like this because I think it'll be less surprising than the gnome-terminal behavior but I'm not that bothered by either choice.  I was previously bothered by the inconsistency, in particular as it applied to the Shell Window tool in Edit, so I was in favor of the previous change at the time.

I don't much like this shared responsibility for parsing the command line, in particular the way I'm not taking advantage of InAppServer's command line parsing, but Shell Window is nice and quick.  I think it's worth the extra lines.