bin/install-everything.sh: Susi and ChrisA both had broken installations due to nightly install-everything's on boxes with make-3.80 dying with:
../salma-hayek/universal.make:556: *** makefile bug: local variable HEADERS from scope setsid (with value "") was referred to in scope setsid.  Stop.

A subsequent sudo install-everything.sh fixed it.  Try to automate that, although I think it's doomed to fail, as I've said in the code.  We unconditionally sudo rm -rf the whole project, source and all, before building, so it's difficult to see how a second attempt to install-everything.sh would succeed where the first one doesn't.  One possibility is a race condition in make.  The only other I've come up with is the one I've suggested in the code comment.