bin/invoke-java.rb: Now there's One True Place for shared libraries to end up, there's no need for terminator to tell us what to put on the shared library path.  There will be a corresponding terminator change along presently.  Note the FIXME: I'm coming back.
bin/make-mac-os-app.rb: I think this change fixes the FIXME here but I haven't tested it beyond building a terminator.dmg.  The comment was quite rusty anyway.  I did notice that trying to rebuild the .dmg fails.  I don't think this was me from the look of the patch.

Building .app bundle for Terminator...
/usr/lib/ruby/1.8/fileutils.rb:714:in `rmdir': Directory not empty - .generated/native/Darwin/terminator/Terminator.app/Contents/Resources/salma-hayek/classes/e/ptextarea (Errno::ENOTEMPTY)
	from /usr/lib/ruby/1.8/fileutils.rb:714:in `remove_dir'
	from /usr/lib/ruby/1.8/fileutils.rb:708:in `remove_dir'
	from /usr/lib/ruby/1.8/fileutils.rb:702:in `foreach'
	from /usr/lib/ruby/1.8/fileutils.rb:702:in `remove_dir'
	from /usr/lib/ruby/1.8/fileutils.rb:708:in `remove_dir'
	from /usr/lib/ruby/1.8/fileutils.rb:702:in `foreach'
	from /usr/lib/ruby/1.8/fileutils.rb:702:in `remove_dir'
	from /usr/lib/ruby/1.8/fileutils.rb:708:in `remove_dir'
	 ... 15 levels...
	from /usr/lib/ruby/1.8/fileutils.rb:655:in `each'
	from /usr/lib/ruby/1.8/fileutils.rb:655:in `rm_r'
	from /usr/lib/ruby/1.8/fileutils.rb:679:in `rm_rf'
	from /private/var/automount/home/martind/software.jessies.org/salma-hayek/bin/make-mac-os-app.rb:45
make: *** [terminator.dmg] Error 1

src/e/util/SpellingChecker.java: We shouldn't need such a contrived path to find the NSSpell executable any more though, again, I haven't tested this.

per-directory.make: Split the idea of GENERATED_DIRECTORY into COMPILATION_DIRECTORY and TARGET_DIRECTORY.  I like the idea, which I've preserved, of compiling into the same directory as the source.  That keeps the make rules simple and traditional: %.o: %.cpp.  I have already, however, overridden the default rule for linking, and I can move the target of that without changing the rule.  Well, beyond adding the idiomatic mkdir for the target directory.  Talking of which, one of the other such mkdirs wasn't idiomatic.

I was confused by the name COMPILED_JNI_HEADER and the need for GENERATED_JNI_DIRECTORY when making the change.  I've simplified that and used names that I think I'll understand more easily in future.

I've had to make all the WiX rules conditional on being "in" the WiX directory, like I already had to do for the JNI.  Otherwise we get warnings from all the other directories about overriding rules for ".msi" etc.  A better solution might be to hive off a wix.make and a jni.make, which are conditionally included.  Another day, perhaps.

While checking that Cygwin built, I realized that a simple "make" in salma-hayek was still building the installer.  That was because the dependency of the default "build" target on the native code was being handled here, rather than via the "native" target, which I edited the other day.  I'd forgotten that the native target was a later addition, to save me time on Cygwin.  Lose that code...

universal.make: ... and defer to the native target.  Now, handily, there's nothing to say that we should even compile the "wixobj" files unless you explicitly ask to "make installer".