per-directory.make: This now says "the JNI header depends on the newly generated version".  Now we're relying on make saying "but it hasn't changed" before then deciding not to build the object file after all.  Previously, it could think that the JNI header was up-to-date, as it had no normal prerequisites, and then try to compile the JNI cpp file before the header was up-to-date.
I suspect that this will cause spurious recompilations on old versions of make, though it didn't the one time I tried it on ted (3.80-9).

I did eventually manage to reproduce Elliott's test case repeatably (3 times wrong without the fix, 3 times right with it).  The trick was to copy with cp -a, to preserve the timestamps.