src/terminator/Terminator.java: car complains that we didn't update our quit behavior when we added the ability to discover what processes are running with access to a particular terminal. So rewrite handleQuit to be equivalent to the user hitting close on all their windows (possibly getting per-window confirmation dialogs), and only exit if there are no windows left open at the end.
src/terminator/TerminatorFrame.java: correct handling of window closing in Java continues to be difficult. WindowListener.windowClosed is wrong because it's only invoked when a window is disposed. WindowListener.windowClosing only applies to closes initiated by the window manager. ComponentListener.componentHidden, though unconventional and not well signposted in the documentation, would be the right choice if it weren't for the unfortunate problem that we're already running on the EDT. So anything such a listener would do would happen too late. Overriding setVisible, it seems, is the only correct alternative in this application (or any application that wants to integrate with Mac OS' standard quit mechanism?). To further complicate matters, setDefaultCloseOperation continues to be misleading, so comment it.

src/terminator/Frames.java: name the hidden frame for ease of debugging. Remove frames from our internal list before iterating over the list. Add a convenience method for copying the internal list (for safe concurrent iteration) because Collections.list is broken.