src/terminator/view/highlight/SelectionHighlighter.java: Implement a nice simple way for the selection highlighter to remove the current selection, based on a region of text which has changed.
src/terminator/view/JTextBuffer.java: Add a method for getting at the selection highlighter, so that...

src/terminator/model/TextBuffer.java: ...the text buffer can notify it when a bunch of lines change.

This isn't perfect, in that editing text above the selection will cause the selection to be cleared, but getting it working properly would involve more work in the TextBuffer/TextLine area.