bin/invoke-java.rb: there's no reason to insist that salma-hayek shares a parent with the project root directory, so remove that restriction.
bin/find-jdk-root.rb: rewrite the subtle copy & paste mistake here to be more obvious in its intent (it's just looking for other scripts in the same directory as it, and doesn't care about projects or salma-hayek), and so that it doesn't modify instance variables in the calling class, whatever that class may be! The run-time lookup of @project_root and @salma_hayek was finding the instance variables in the class called "Java" from "invoke-java.rb". Lexical scoping continues to be evil.

With these changes, a project's start-up script needs to know where salma-hayek is (so that it can require "invoke-java.rb") but that can be anywhere no the file system, and everything will just work. The only other restriction I'm aware of is that your start-up script must be in the "bin/" directory off your project's root directory.