universal.make: Follow the convention of only using immediate assignment when there's a particular reason.
One common reason is that we're looking at the file system and only want to do this once, before rules are run.  Another is that we're invoking an external command.  Both of these are performance rather than correctness concerns.  Most external commands and file system inspections are quick but some aren't and, rather than evaluate them (ie guess) on a case-by-case basis, it's easier to just adopt a convention.

There are other reasons for using := and they occur in the kind of code you don't want to change carelessly.  Which is say that := sometimes means "danger here".