TODO: I can never remember "socat" when I want it.  It'd be worth writing up for the web page.  Elliott suggests that we'll look like stoners if we write it up before trying it.  So I'll add it to TODO, the home of half-baked ideas.
The last person that I remember wanting a serial back-end (simonj) was really wanting to replace Windows' default-installed HyperTerminal but hadn't (and, afaik, still hasn't) found anything else with the same quality of progress reporting.  That's not something which ought to be part of terminator (like telnet and ssh) but it would be eminently reasonable to recommend another program on the terminator web page.