src/terminator/model/TextBuffer.java: Why would we want to insert a line when we move to the last existing line?  Surely we only want to insert if we move one down from there.  This fixes readline editing near the start of a once-wrapped line at the bottom of the window (so a line between 81 and 160 characters in a default-sized terminal).  Previously, the part of the wrapped line on the last line of the window would get cleared.
Although I'm not very confident in it, this change doesn't break scrolling down in vim for me, even in the split-screen :help mode.  My first attempt changed the condition from being on the scroll index to being on the size of textLines and that broke vim.

All these months suffering that.  I hadn't realized that it was our fault.