bin/find-jdk-root.rb: improve comments.
universal.make: simplify and clarify the stuff related to invoking javac(1), with an eye to moving it into "invoke-java.rb". Dump the non-functional support for gcjx(1) because the only way OSS Java is going to catch up with Sun's real, working, free Java is if Sun stops development, at which point we'll have greater worries than our build system.