bin/annotate-patch.rb: bin/blogger-backup.rb: bin/dictionary: bin/download-latest-jdk.rb: bin/ensure-suitable-mac-os-version.rb: bin/escape-for-html.rb: bin/fatbits: bin/file-list-to-wxi.rb: bin/invoke-java.rb: bin/javahpp.rb: bin/mailpatch.rb: bin/make-img-tag.rb: bin/make-mac-os-app.rb: bin/make-version-string.rb: bin/nightly-build.rb: bin/notepad: bin/scm-diff.rb: bin/send-meeting-requests-to-ical.rb: bin/svn-log-to-html.rb: bin/uuid.rb: bin/which.rb: bin/xv: escaping.rb: patch-to-html-email.rb: Change all the Ruby scripts to be executed via /usr/bin/env, so they work wherever ruby is on the path, even if not in /usr/bin.
universal.make: With this change, the www-dist and source-dist targets should work on crank.org.uk without the explicit "ruby " here.

This is probably the best place to record the rationale.  From Chris's Solaris sysadmin perspective, it's a pain to put ruby in /usr/bin.  I'm not sure whether this one file is really a pain or whether he regards it as the thin end of a wedge.

I don't think it is the thin end of the wedge.  Un*x hash-bang lines mean that script interpreters are usually assumed to be in /usr/bin.  Script interpreters are thus a special case.  Even though this env trick looks like a good workaround for us, we might run into pain later.  Imagine you're a user who downloads some program which includes some scripts.  Those scripts are likely to refer to /usr/bin/ruby.

The env trick may cause us problems on some client.  But, if we rely on those "ruby " prefixes in universal.make, the time *will* come when we run another post-commit script and forget the prefix.  post-commit will break and we'll be baffled again.  So let's avoid the problem we know is coming.

Elliott [

actually, i see that neither the ruby nor the python man page on Mac
OS refer to the env trick. so maybe it's just in the documentation.
(also, both man pages refer to the language's default location being
/usr/local/ despite the fact that i've never seen it ship in that
location. i guess they don't care enough to have the makefile rewrite
the man page when they install it. or wrote all this before they were
popular, and haven't noticed it's out of date. it's not like i'm about
to bother filing bug reports, after all.)

the Perl man page gives a clue to why the env trick isn't more widely
used; the implication being that you can't trust that you (a) have
env(1) or (b) know where it is. (someone should have a web site with a
big table showing such things. does POSIX specify locations for
standard executables?)

Google suggests the following order of popularity:

/usr/bin/ruby
/usr/bin/env ruby
/usr/local/bin/ruby

"env ruby" finds a few "/bin/env ruby" lines, but not enough to look
like they're not mistakes, or to affect the results above.

env(1) would appear to have the advantage of letting people run a
custom interpreter on a system with a default interpreter, too. i've
just always assumed that i'm missing some big reason why you don't
want to use env(1), because ever since i first read perlrun(1) i've
thought it looks like the best choice.

maybe it's the performance concern of the extra exec?

], Martin [

And the extra path search.  Exposing yourself to the vagaries of the user's path would be another reason.  Not needing it would be the killer.  I don't remember the last time I saw a Perl script which didn't start #!/usr/bin/perl.

].