native/all/launcher/launcher.cpp: Call out to some code to fix the Windows native environment when running under Cygwin.
native/all/launcher/setup_win_environ.h: Declare said code.

native/all/launcher/setup_win_environ.cpp: Rip-off said code wholesale from Michael Schaap's GPL cygstart.c.  We're GPL too, so we can do that, woo.

Why do we need to do this?  It's a long detective story...

We've known for a while that we can't get at $HOME from Terminator on Cygwin.  Since we've had the Show Environment item on the Debug menu, we've been able to see that the environment is very limited indeed.  This caused Brian Hobbs some grief today when we kept starting /bin/sh instead of /bin/bash, despite his having set $SHELL.

After studying lots of interweb (MSDN and more dubious sites) and objdump -p output, I came to the conclusion that Win32 Java probably implements System.getenv via the Win32 GetEnvironmentStrings API.

A small C++ program to dump out the returned memory from GetEnvironmentStrings showed that, when compiled as a mingw32 program, the environment was large (ie as we'd hope) but, when compiled as a Cygwin program, it was as constrained as seen from Terminator's Java.

I found the code in spawn.cc which passes real_path.iscygexec()
(see how-spawn-works.txt) to build_env's no_envblock parameter, meaning
that, when starting a cygwin executable, it doesn't pass the Windows
native environment to CreateProcess.  I've looked at all the calls to
SetEnvironmentVariable - which is the other way (the only other one I
know) of altering the native environment - and that's never used to
populate the whole native environment, just the odd variable here and
there.

There is a call cygwin_internal(CW_SYNC_WINENV) which would let the
application decide to synchronize the native environment from the Cygwin
environment (doing path format translations where appropriate) but that
was only added a few weeks after the last release.

The reason for this deviousness turned out, I think, to be a limit in the
amount of environment you can pass via CreateProcess:

http://sourceware.org/ml/cygwin/2005-05/msg00081.html 

If the parent and its Cygwin child collude, they can pass over an
environment of any size by some other means.  In the case of running our
launcher, they can be sure the child is a Cygwin process, so this must
seem like it's bound to work.  If it's bound to work, there'd be no
point in building up an environment block to pass to CreateProcess to
set the Windows native environment in the hope it's less than 32KiB (or
whatever the limit is).  We're a whacky special case in that we're a
Cygwin process but we contain a non-Cygwin DLL that wants to read the
(native) environment.

Roughly the same issue recently caused a problem for Cygwin's cygutils' cygstart.  There's some hint in the emails that some 1.5.19 change made this more problematical.  (cygstart is a Cygwin program which starts its usually-native child via ShellExecute rather than fork-and-exec, which is why it's also a whacky special case.)  The problem was batted around the usual suspects, resulting in a patch here, that comprises most of the code I've copied:

http://www.cygwin.com/ml/cygwin/2006-02/msg00321.html

I followed more false leads than I would ever have imagined while trying to find the current home of that program, so I've left a URL in the source.

It's not clear from the ChangeLog or my google searching, whether the CW_SYNC_WINENV is intended to tidy this up.  Like many projects out there, the ChangeLog concentrates on what changed, not why.