universal.make: paulb couldn't start Edit this morning. he'd already permanently deleted his build mail. install-everything, run again, said that edit.tgz was the same as on disk. tar reported trailing junk, and the script stopped at that point.
The corrupt edit.tgz contains all the same bytes as the previous revision from the one which should have been (and probably was) extant on the server at the time of Paul's cron job.  It also contains the last 25 bytes of the newer revision, making it the same length as the newer revision.  I'm having trouble imagining how that could happen, whether on the client or server side.

If the .tgz file on the server gets replaced while it's being served, it's seems quite possible that we'll download a truncated version, even if we were to ensure that the replacement is atomic.  Still, it only seems like a good idea to make the replacement atomic.

Before we used rm -f && cp, we used cp -f, which wasn't portable when the target was a symlink.  Before we used cp -f, we used ln -s, which was why some of the targets were symlinks.  Now I believe that the permissions on the directories and the umask under which this job runs are right.  So I think we should be OK with mv.  You could well argue that this isn't The Make Way and that, if we're going to behave like this, we should just do this unconditionally with a script.  I never did like leaving .tgz files lying around in /home/software/checked-out.