src/e/edit/NewFileAction.java: Move a special case...
src/e/edit/Workspace.java: ... to where it will be of more use.  This means that Open (as opposed to Open Quickly) of a file that Edit hasn't seen will cause the workspace index to be updated.  As will opening a file via the EditServer, though Chris didn't report that.  It does nothing to help the Save As case (or the Save As semantics) which Chris also reported.