universal.make: add a rule to create a ".deb" package.
bin/package-for-distribution.rb: make the Mac OS-specific stuff optional and add in some Linux-specific stuff. It's all done procedurally, but there aren't actually all that many differences. Note that the created ".deb" packages are not yet suitable for use. We need to look in to whether we can use our current combined version "number" instead of the hard-coded version number in this script, I need to integrate my (working) ".desktop" stuff, we need to fix the terminfo situation, and the project's start-up script(s) need to be in the fake /usr/bin/.