src/terminator/view/highlight/SelectionHighlighter.java: Borrowing heavily from PTextArea's implementation of mouse selection, re-implement Terminator's selection code in a better way.  Double- and triple-click drags will now select and deselect in multiples of words or lines.
It's rather unfortunate there's no easy way of moving this code into something which can be used by both PTA and Terminator.  The principle is shared, but the interface to the underlying selectable thing would be so large (particularly for the PTA version) that I can't see it as being worthwhile.

TODO: Remove done job.