src/e/util/IPC.java: IPC contains new abstraction interfaces for servers which generate connections to things.  A connection is an input and an output stream.  IPC also has an implementation which uses a ServerSocket and Sockets, and one which relies on FIFOs.
src/e/util/InAppServer.java: Modify the InAppServer to attempt to use the FIFO mechanism, and if that fails, use the old Socket implementation instead.

Note that the format of the little file which stores the connection information has changed, so that it's prefixed with 'socket:' for the old stuff, and is of the form 'fifo:<filename>' when using FIFOs.  This allows the ruby scripts to determine which connection mechanism to use.