src/terminator/view/JTerminalPane.java: stuartb complained that he was seeing the visual bell too often. Specifically:
  In vi, already in command mode, hit escape - I see two "bells" in terminator, one in xterm.
  In insert mode, hit escape - one bell in terminator, none in xterm.
 
  At bash command line, hit escape twice [quickly] - both terminator and xterm say "Display all <x> possibilities? (y or n)", but only terminator bells.

I wondered if it was the work-around for Sun bug 6320676 causing a problem on Java 5, but I can repeat the problem with Java 6 too. Don't doubt Mr Bell! Two lines after a comment saying "don't handle keys that generate KEY_TYPED events", we were handling VK_ESCAPE, which generates a KEY_TYPED event. Stop doing that, and the Escape key works as it should.