lib/installer.wxs: A copy (with no changes) of terminator.wxs, which hasn't (since its last change) contained anything terminator-specific.
I suspect Subversion's going to cause me a problem now, by committing some other changes that I haven't selected.  I need to check in this file before the others because I'm going to have to change this file to refer to the new location of the .ico file.  I want to do that in a separate check-in, so the history shows the diff.

lib: Yeah, it won't let me commit unless I include this directory.