per-directory.make: Lose the definitions of (BIN|LIB)_DIRECTORY which weren't directory-specific anyhow to universal.make.  Lose all the WiX code there too.  I need the rule for copying source there too, but I'll still need it here, so abstract it out.
universal.make: Now all of the installer-generation is done here, it's clearer how we should specify the installer filename in a per-OS way.  Two of the FIXMEs, regarding the "native" target, were inconsistent with our carefully crafted Cygwin behavior.  "make native" doesn't build the installer there, so you can still try building the native code without the WiX tools installed.  "make installer" should build the .dmg or the .deb or the .msi.  That's easily implemented and should make the native-dist target "just work" on Darwin.

For some reason, the native-dist target was uploading terminator.msi as terminator-Cygwin.msi.  That's inconsistent with app-dist's upload of terminator.dmg.  terminator.msi will be fine.

I tried to reduce the clutter of over-generality in the WiX variables by moving some of the code into the rules.

lib/installer.wxs: I've moved the .ico file to lib/ alongside the .png files from which it was manually generated via the gimp.  Now the WiX installer's no longer generated out of the native/ tree, it doesn't make sense to generate its files in .generated/native/.

native/Cygwin/WiX/salma-hayek.wxs: Remove the old copy of module.wxs, which Subversion kindly half moved in the last check-in.

native/Cygwin/WiX: And the directory it rode in on.