native/Cygwin/WiX/terminator.wxs: terminator.wxs has moved to salma-hayek/lib/installer.wxs...
native/Cygwin/WiX: ... removing the reason for this directory.

native/Cygwin/WiX/terminator.ico: 
lib/terminator.ico: Move the .ico alongside the .png files it was created from, now there's now WiX directory for it to live in.

www/index.html: The native-dist rule has changed to remove the spurious -Cygwin in the uploaded installer name, for consistency with the old app-dist rule.