native/all/pty/terminator_terminal_PtyProcess.cpp: Thomas Duffey has been reporting for some time that if he runs vim(1) in the third or fourth tab in a window on Mac OS, its idea of the terminal size is wrong. We'd been unable to reproduce it and hadn't got very far. Today, Simon Pamies reported the same problem in almost exactly the same words, but in a way that made Martin Dorey wonder about how to write a test program and then have the idea of inserting a call to sleep(3) in "PtyGenerator.h" to see if we could trigger the problem. We could, and it turned out to be that if we caused the child to sleep before opening the pty, instead of the TIOCSWINSZ ioctl blocking as it does on Linux, Mac OS seems to throw the window size information away.
We talked about a variety of solutions. We didn't like the idea of a separate pipe to communicate between parent and child (used by some of the competition) because it seems unnecessary; we didn't like the idea of using the traffic that will eventually come from the child because a quiet child could cause us to block far longer than necessary; we didn't like the idea of having the child write something for its parent before calling execvp(3) because it requires extra code in both parent and child. We settled on the idea of using select(2) to recognize when the child was ready (on both Linux and Mac OS, writes -- as opposed to ioctls -- block if the child hasn't opened the pty). This requires relatively little code, and only in the parent, because it synchronizes on something that the child has to do anyway. This doesn't appear to have had any negative effect on Cygwin, Linux, or Mac OS, and has fixed our artificially-induced problem on Mac OS.