universal.make: stop passing the individual .class files for packaging.
bin/package-for-distribution.rb: generate a combined uncompressed JAR file of both the project's classes and the library classes. This gives us the fewest files to install and -- more importantly -- the fastest start-up time. Also a variety of fixes so that the Mac OS build can be done on and/or installed on a case-sensitive file system. Pull some of the previously Linux-specific permissions-fiddling code out so it can be shared with Mac OS, so that a Mac OS build done on a file system with setgid directories can actually be installed (otherwise the Finder complains that "One or more itmes have special permissions and cannot be copied.", an error which doesn't seem well understood around the net). Use the "human" project name in the .deb control file's description field, since that's meant for humans.

src/e/gui/AboutBox.java: File.exists doesn't seem to like ".." in paths starting with non-existent directories, but it will let us use File.getParent; we need this because...

bin/invoke-java.rb: ...the classpath now has two alternatives, one for end-users and the other for developers. Also another part of the fix for Mac OS builds that work on case-sensitive file systems.