bin/install-everything.sh: Fix the BlueArc nightly installations.
parent.make: Fix my daily build.

universal.make: Re-add the classes files to the Win32 installer.  When package-for-distribution.rb can distribute everything and we have a strategy for deciding which files from which copy of salma-hayek's bin go on the user's PATH, then it'll be high time to switch over the Win32 installer to use the same directory structure - and hence the .jar file.  I could switch it over sooner but I think it'd cause more breakage than it'd avoid.