universal.make: Pass the salma-hayek installer file list to the packaging script.  It involves one spurious make invocation (and one necessary one) because the Win32 installer doesn't want the $(PROJECT_NAME)/ prefix and would have to jump through an unpleasant hoop to remove it.
bin/package-for-distribution.rb: But let's us fix a FIXME here and get rid of a large wedge of code.  The only disadvantage that I've seen when building terminator.deb is that we now include salma_hayek/.generated/build-revision.txt - a pointless file, unless we're installing a separate salma-hayek package.  We're already including in the Win32 installer, so that wants fixing in the make rules.