src/terminator/view/JTerminalPane.java: Lose responsibility for selecting the default shell...
bin/terminator: ... to somewhere that we can test whether /bin/bash is executable on Cygwin.  We can't do that directly from the Java because "/bin" only has meaning to Cygwin, not to Windows-native code like the JVM.

This change means that we're setting an environment variable in a Cygwin application (rubyw.exe) and relying on being able to read it from a Windows-native DLL (jvm.dll) in another Cygwin application (launcherw.exe).  This wouldn't work but for synchronizeWindowsEnvironment - the problem we previously had with $HOME.  That code looks quite dubious and the next version of Cygwin will obsolete it (although we won't necessarily want to remove it immediately).  So this change might have increased our fragility.  It'll be useful to know that it's broken, rather than masking the problem, as we previously did.

This gives us a better Cygwin out-of-the-box experience.  Cygwin users are used to clicking on a shortcut to Cygwin.bat, which runs bash --login -i.  When run as bash, the shell gives you a different default prompt than if run, as we were running it before, as sh.  The "posix" flag (which is what running it as sh sets) interacts with the "vi" line-editing mode flag to disable tab completion, as discovered by Chris Oklota.

On other platforms, SHELL is probably already set, usually to /bin/bash, and this change should have no functional effect.  If it isn't set, we might run bash where before we ran sh.