post-commit: The reason we weren't getting told why the post-commit script was failed was because, on this system, "mail" is exim and:
exim abandoned: unknown, malformed, or incomplete option -s

We already use sendmail directly in the patch-to-html-email.rb script which we call.  It's a bit distressing not to abstract-out the duplication.  I guess the idea of using "mail" was that it was simple and should Just Work.  Chris probably has a more useful opinion on what we should be doing.