src/e/edit/ECloseButton.java: src/e/edit/ESwitchButton.java: defer the default foreground color choice to EButton...
src/e/edit/EButton.java: ...where we in turn defer it to our component hierarchy parent...

src/e/edit/ETitleBar.java: ...and set the foreground color for the buttons' JPanel to be the same as the color we're using for the title bar text. This makes the buttons visible on Windows, where they're otherwise black on dark blue. Also change the title bar font on Windows to match the window title font, to aid its visibility. Stick to the table header font on the other platforms, because we don't need such a heavy font there, and it's just distracting.