bin/install-everything.sh: The ithaki build failed partly because I hadn't installed alien there.  I re-ran it but the Evergreen rpm is still missing because of a case issue but, before I forget this, make a note.
Apropos of doing this in a more appropriate place, this code has been misleading Stuart into thinking his build server is being updated every night.  The trouble with "apt-get install" is it also means "upgrade" and we're not really trying to upgrade here.  installMissingPackages is a gruesome hack too.  There must be a better way.