bin/humanize.rb: As Elliott pointed out, how would you humanize "knowall"?  (Well, you'd probably start from "know-all", but the point is that we should shouldn't try to reverse potentially lossy transformations.)
bin/invoke-java.rb: My search for "downcase" found a hit here.  This is only used from "scm" at the moment, which has "scm.icns" and "scm-128.png", so the questions here are academic at the moment.  If and when we have a non-trivial use, we'll have to fix this.

lib/installer.wxs: Change the .ico files to MACHINE_PROJECT_NAME, as previously mentioned.  Elliott says "lower case in URLs is the clear winner" which is fortunate because it would be difficult to preserve the previous behavior.  I've renamed the Evergreen downloads directory on the server appropriately.  The shortcut and the package name (as shown in Add/Remove Programs) changes to the human name, which is an unforeseen improvement for Terminator.

bin/package-for-distribution.rb: universal.make needs to know about human and machine names.  We could pull it out into a script but, at the moment, it's easier to just pass both names in.  I've used the machine_project_name almost everywhere we previously used the project_name (as opposed to the human_project_name).  This keeps Terminator's installer unchanged but changes "Evergreen" to "evergreen" in many places in Evergreen's.  The "almost everywhere" exception is in the case of the .icns file.  Elliott thinks that we might have a problem if that has a difference case to the package title.

universal.make: It would be nice to have the installer's contents independent of the name of the directory from where it was built.  Distributed version control systems encourage people to have multiple work-areas, often called things like "terminator.line-wrap".  That's one, medium term motivation.  The immediate motivation is to make it such that universal.make can work out the name of Evergreen's rpm, rather than being tripped up when alien(1) downcases the name.

At the moment, we're defaulting the HUMAN_PROJECT_NAME to the name of the directory.  I already overrode that for Terminator.  We could for the other projects, if this proves a successful experiment, though I expect we'll leave it until we need to.

As previously mentioned, the thought is that filenames should be lowercase unless there's a reason otherwise.  So the downloads directory on the server becomes lowercase, the distribution files all become lowercase, the jar file becomes lowercase and various phony targets become lowercase.

One change I hadn't foreseen when I started out was the occasional need for the project directory's name.  Sometimes this is the same as the machine project name (in the case of "terminator") and sometimes it's the same as the human project name (in the case of "Evergreen").  Now we're explicit about which name we want at the point of use, we might break things less often.

I've made some effort to build, test and upload the .deb and .msi installers.  The .rpm installers are built and uploaded.  I expect the .dmg ones are subtly broken, though I have tried to be careful.