bin/package-for-distribution.rb: Lose the code to install the .deb building packages.
bin/install-everything.sh: Lose the code to install prerequisite Debian packages, both for running and building.  I realized I needed to update the KnowAll download directory on the server.  And, oh god, the directory inside the source-dist.tgz file doesn't have the same capitalization as the file we now download.  All for the sake of keeping that part of this script alive for one more day.

bin/install-build-prerequisites.sh: Gain all the package installation code.  The vision here is that the salma-hayek build instructions will change to say "run this file once as root or do the equivalent for your system".