universal.make: Save Ed from having to guess what today's .rpm is called by making it available as:
http://software.jessies.org/terminator/org.jessies.terminator.rpm

alien is, if not forcing this on us, then encouraging us to do this (perhaps with good reason).  I think it would be useful for .deb packages because it would make it more likely that the package that apt-get was expecting to find was still there.

The easiest way to achieve that would be to do it for all package types.  But do we really want org.jessies.terminator.dmg?