universal.make: Desist from passing the file list to the packaging script on stdin...
bin/package-for-distribution.rb: ... in favor of having the packaging script ask the makefile for the list.  By jumping through this hoop, the packaging script knows whether it's getting a list that should go in the project's directory or in salma-hayek's.  This means that it doesn't have to know how to rewrite "Evergreen.feature/..." as "evergreen".

You might well think "why don't we generate the file list in Ruby instead?".  As commented in the makefile, on the line which generates the file list, it's only the makefile which knows about all the per-directory targets (ie native executables) in a way that's robust against old cruft that may be lying around.

This gets rid of a nasty wart in the makefile without polluting the packaging script with flaky path rewriting hacks.  Oh yes, and means that evergreen.deb will install /usr/bin/evergreen again - which has been broken since my change to make the difference between human_project_name and machine_project_name a more explicit one.