src/e/gui/RepeatingComponentTimer.java: I could readily repeat the CursorBlinker-timer-keeps-Terminator-alive-when-it-shouldn't bug this afternoon (doing runtitan --connect titanusstress1).  HungAwtExit told me it was that timer.  Printf debugging told me that JTextBuffer was being told focusGained() after the RepeatingComponentTimer had been told that the component was no longer displayable.  focusGained() was thus restarting the timer after it had been stopped.  Who knows why, and who cares - it's probably not our fault.  We can guard against this by not starting the timer when the component's not displayable.  We need to keep track of the caller's intention so that, should the component subsequently become displayable, we'll start the timer then.
I think "armed" was Elliott's suggestion for the name of the necessary stored state.  I'm not fond of changeTimerState either - it carries too much of a toggling connotation for my liking.  "Normalize" would be my choice of verb in my code and it does seem to get a lot of mentions in the JDK source, though I haven't looked at any of them to see if the connotation is right.