lib/installer.wxs: Add ourselves to the end of the PATH for all users on the system, for the benefit of scm.  The uninstaller correctly removed the PATH entry for the one, simple case I tested.  (I couldn't face looking at what it does if you've edited the component of the path in question.)
I could perhaps have added the Environment tag to another Component but that doesn't seem the WiX way, so I added a new component.

The difficult part was finding that this perturbed evaluation order of the CYGWIN and RUBYW properties such that, even though the RUBYW property uses the CYGWIN property, it was now evaluated first - and thus didn't work.  What a half-baked little language.

universal.make: Provide a guid for the new component.